tts/requirements.txt
aiohttp==3.9.1
piper-tts==1.2.0
numpy==1.24.3
